API/requirements.txt
opencv-python==4.11.0.86
pandas==2.2.3
pillow==11.2.1
scikit-learn==1.6.1
setuptools==80.4.0
fastapi==0.115.12
uvicorn==0.34.2
tensorflow==2.19.0
numpy==2.1.3
pydantic==2.11.4
pydantic_core==2.33.2
python-multipart==0.0.20
requests==2.32.3
